Find DB table name using regex search.

Some Rgemini functions internally query DB tables. The table names cannot be hard-coded in those functions since HPC datacuts sometimes have slightly different table names (e.g., admdad is called admdad_subset in some datacuts). This function uses a simple regex search to identify the full table name in a given DB that matches the DRM (Data Reference Model) table name of interest.

Currently, the function only supports a subset of table names (see below) and expects the relevant tables in all databases to only differ based on their suffix (e.g., "admdad" vs. "admdad_subset"). Specifically, the function currently uses grepl("^tablename",drm_table) to look for table names that start with the same name as specified in DRM (e.g., 'admdad').

Usage

find_db_tablename(db, drm_table, verbose = TRUE)

Arguments

db

(DBIConnection)
RPostgres DB connection

drm_table

(character)
Table name to be searched, based on the DRM. Currently only accepts the following inputs (which have been verified to work across different DBs/datacuts):

"admdad"

"ipdiagnosis"

"ipintervention"

"ipcmg"

Users need to specify the full DRM table name (e.g., "admdad" instead of "adm") to avoid potential confusion with other tables.

verbose

(logical)
Whether or not to show a message indicating which DB table was found.

Value

(character)

Returns the full name of the relevant DB table as a character.

Examples

if (FALSE) {
drv <- dbDriver("PostgreSQL")
db <- DBI::dbConnect(drv,
  dbname = "DB_name",
  host = "172.XX.XX.XXX",
  port = 1234,
  user = getPass("Enter user:"),
  password = getPass("Enter Password:")
)

admdad_name <- find_db_tablename(db, "admdad")

# query identified table
admdad <- dbGetQuery(db, paste0("select * from ", admdad_name, ";"))
}