Compute frailty score using ICD-10-CA

frailty_score returns frailty risk score for hospital admissions calculated based on ICD-10-CA diagnoses codes.

Usage

frailty_score(cohort, ipdiagnosis, eddiagnosis)

Arguments

cohort

(data.frame or data.table) Cohort table with all relevant encounters of interest, where each row corresponds to a single encounter. Must contain GEMINI Encounter ID (genc_id).

ipdiagnosis

(data.table) ipdiagnosis table as defined in the GEMINI Data Repository Dictionary. This table must contain the genc_id and diagnosis_code (as ICD-10-CA) fields in long format table only. The diagnosis codes must be free from any punctuation or special characters.

eddiagnosis

(data.table) eddiagnosis table as defined in the GEMINI Data Repository Dictionary. This table must contain the genc_id and er_diagnosis_code (as ICD-10-CA) fields in long format table only. The diagnosis codes must be free from any punctuation or special characters.

Value

data.table

This function returns a table with all encounters identified by the cohort table input and additional derived numeric fields for ip_frailty_score_derived, er_frailty_score_derived

and frailty_score_derived

ip_frailty_score_derived : Frailty risk score calculated only using IP diagnoses codes from ipdiagnosis table

er_frailty_score_derived : Frailty risk score calculated only using ED diagnoses codes from eddiagnosis table

frailty_score_derived : Frailty risk score calculated using both IP & ED diagnoses codes

If frailty risk score is < 5, the patient is low-risk If frailty risk score is between 5 and 15, the patient is intermediate risk If frailty risk score is > 15.0, the patient is high risk

Details

This function takes a list of ICD-10-CA and generates a numeric field corresponding to frailty risk score for each admission. This function can take any data.table with ICD-10-CA codes. Historically, scores have been calculated based on partition of ICD-10-CA codes.

Researchers may want to focus on frailty risk scores based on in-patient (IP) diagnoses because emergency department (ED) diagnoses have lower sensitivity/specificity for certain conditions and are less reliably filled compared to IP diagnoses. In general, there will be overlap between IP and ED diagnoses codes. ED diagnoses codes to compute frailty risk score can be used to evaluate frailty risk upon admission to IP services.

This function does not differentiate between diagnosis types.

Note

The original frailty risk score was developed based on ICD-10 codes instead of ICD-10-CA. However, ICD-10-CA version codes are identical to the ICD-10 counterpart up to 4 character levels and the frailty risk score only uses the first three characters of ICD-10.

Please refer to the references in this page for more details.

Warning

This function returns a data.table with genc_id and three numeric fields. The function will return NA values if genc_id in the cohort table is not found in the ipdiagnosis or eddiagnosis tables.

frailty_score_derived is NA if either ip_frailty_score_derived or er_frailty_score_derived are NA. When one tries to left-join the output of this function with another table (another list of admissions in the left), make sure list of admissions (or patient ids) aligns in both tables.

References

Gilbert T, Neuburger J, Kraindler J, et al. Development and validation of a Hospital Frailty Risk Score focusing on older people in acute care settings using electronic hospital records: an observational study. Lancet 2018; published online April 26. http://dx.doi.org/10.1016/S0140-6736(18)30668-8.