Compute the number of radiology tests per encounter

n_imaging returns the number of radiology tests for hospital admission.

Usage

n_imaging(
  cohort,
  imaging,
  dtvar = "performed_date_time",
  mapvar = "modality_mapped"
)

Arguments

cohort

(data.frame or data.table) Cohort table with all relevant encounters of interest, where each row corresponds to a single encounter. Must contain GEMINI Encounter ID (genc_id).

imaging

(data.table or data.frame)
Table equivalent to DRM table "radiology" as defined in the GEMINI Data Repository Dictionary. Table must contain three fields: genc_id, imaging performed date-time (see dtvar, in "yyyy-mm-dd hh:mm" format), and a field that standardizes the imaging test type (see mapvar).

dtvar

(character) Name of the column in imaging table containing date-time of performed imaging test (usually "performed_date_time").

mapvar

(character) Name of the column in imaging table containing the imaging test type (usually "modality_mapped").

Value

data.table with the same number of rows as input "cohort", with additional derived numeric fields labelled as "n_img_xray_derived", "n_img_ct_derived", "n_img_mri_derived", "n_img_us_derived", "n_img_other_derived", "n_img_int_derived" and "n_img_ct_mri_us_derived".

Details

This function takes a list of admissions and DRM table equivalent of imaging table to generate numeric fields counting the number of different radiology tests for each admission.

Imaging table in the database should include field that classifies each test into below 7 categories.

1 : X-Ray

2 : CT

3 : MRI

4 : Ultrasound

5 : Other

6 : Interventional radiology procedures

7 : Echo (Excluded)

Therefore, this function should be run after some standardization efforts on imaging table (performed by GEMINI team). Currently, the number of advanced radiology tests is one of the performance metrics in MyPracticeReport. CT, MRI and Ultrasound are considered advance imaging tests, which can be retrieved by the derived variable (n_img_ct_mri_us_derived).

Any tests being performed in both ED and in-patient settings will be counted.

Warning

Function returns data.table with id field and several numeric fields. By design, function will not return any NA values. When one tries to left-join the output of this function with another table (another list of admissions in the left), make sure list of admissions (or patient) aligns in both tables.

Note

The function takes optional inputs specifying the field label for the variables corresponding to date-time of the performed imaging test and mapped imaging modality, respectively. This is to avoid hardcoding variable names within the function in case field labels differ between databases.