Compute the number of routine bloodwork tests per encounter

n_routine_bloodwork returns the number of routine bloodwork (Sodium and Hemoglobin tests) for each hospital admission. Sodium and Heamoglobin tests are defined by OMOP codes. Sodium is the code 3019550. Hemoglobin is 3000963.

Usage

n_routine_bloodwork(dbcon, cohort)

Arguments

dbcon

(DBIConnection)
A database connection to any GEMINI database.

cohort

(data.frame or data.table) Cohort table with all relevant encounters of interest, where each row corresponds to a single encounter. Must contain GEMINI Encounter ID (genc_id).

Value

data.table object with the same number of rows as input "cohort", with additional derived numeric field labelled as "n_routine_bloodwork_derived"

Details

This function takes a list of admissions and DRM table equivalent of lab table (see GEMINI Data Repository Dictionary) to generate numeric fields counting the number of Sodium and Hemoglobin tests for each admission.

Lab table in the database should include field that classifies each test into a few different categories. Therefore, this function should be run after some standardization efforts on lab table (performed by GEMINI). Currently, the number of routine bloodwork tests is one of the performance metrics in MyPracticeReport.

Function automatically removes tests without valid numeric result value. Exceptions are boundary result value (e.g. ">120" mmol/L), and these are still counted as valid tests.

Any tests being performed in both ED and in-patient settings will be counted.

Warning

Function returns data.table with id field genc_id and one numeric field indicating the number of bloodwork tests per encounter. By design, function will not return any NA values. If a genc_id does not have any entries in the "lab" table, the admission gets assigned 0 number of tests. When one tries to left-join the output of this function with another table (another list of admissions in the left), make sure list of admissions (or patient) aligns in both tables

This function requires mappings by a Subject Matter Expert to ensure that all tests are mapped to 3019550 - Sodium (Moles/volume) in Serum or Plasma, and 3000963 - Hemoglobin (Mass/volume) in Blood